# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel perfume api",
        "description": "L5 Swagger OpenApi ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/admin/search": {
            "get": {
                "summary": "گرفتن محصولات موجود با توجه به فیلتر برای ادمین(ادمین کالا های حذف شده هم میبیند)",
                "description": "category[eq][0]=woody&category[eq][1]=orbal هر پارامتر فیلتر میتواند چند بار تکرار شود مثلا با تکرار کردن فیلتر کنگوری هر دو کنگوری سرچ شده به نتایج اضافه میشوند، برای هر فیلد تکراری کلید ارایه بعد جدید استفاده کنید.",
                "operationId": "6653b04aa6aed09cbcecf71f7d7dab86",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "path",
                        "description": "احراز هویت با توکن",
                        "required": true
                    },
                    {
                        "name": "name[eq]",
                        "in": "query",
                        "description": "فیلتر بر اساس نام",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "quantity[gt] or quantity[eq] or quantity[lt]",
                        "in": "query",
                        "description": "فیلتر بر اساس تعداد",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price[gt] or price[eq] or price[lt]",
                        "in": "query",
                        "description": "فیلتر بر اساس قیمت",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "volume[gt] or volume[eq] or volume[lt]",
                        "in": "query",
                        "description": "فیلتر بر اساس حجم عطر",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "warranty[eq]",
                        "in": "query",
                        "description": "فیلتر بر اساس گارانتی داشتن یا نداشتن (در صورت پر کردن فیلد گارانتی، فقط محصولات با گارانتی به نتایج می‌آیند، برای گرفتن همه محصولات این فیلد را خالی بگذارید)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "gender[eq]",
                        "in": "query",
                        "description": "فیلتر بر اساس جنسیت عطر",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "brand[eq]",
                        "in": "query",
                        "description": "فیلتر بر اساس برند",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category[eq]",
                        "in": "query",
                        "description": "فیلتر بر اساس کتگوری",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "priceAsc",
                        "in": "query",
                        "description": "رده بندی بر اساس گران ترین",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "priceDesc",
                        "in": "query",
                        "description": "رده بندی بر اساس ارزان ترین",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "newest",
                        "in": "query",
                        "description": "رده بندی بر اساس جدید ترین",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "sold",
                        "in": "query",
                        "description": "رده بندی بر اساس پر فروش ترین",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful search",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/searchAdmin"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/perfume/{product:slug}": {
            "get": {
                "summary": "گرفتن اطلاعات کامل مربوط محصول خاص برای ادمین(قابلیت دیدن محصولات حذف شده یا غیر فعال)",
                "description": "اسلاگ مربوط به برند در یو ار ال شود",
                "operationId": "f578ce3135b37fc197c693ddc92e70c0",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "path",
                        "description": "احراز هویت با توکن",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fullProduct"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            }
        },
        "/api/admin/perfume/based-factor/{slug}": {
            "get": {
                "summary": "ادمین قابلیت این را دارد که تمام محصولات از فاکتور های مختلف که محصول ورودی را ارایه میکنند را ببیند",
                "operationId": "07ad82631c8759f03983603f1c780d27",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "path",
                        "description": "احراز هویت با توکن",
                        "required": true
                    },
                    {
                        "name": "اسلاگ مربوطه",
                        "in": "path",
                        "description": "اسلاگ مربوط به محصول در یو ار ال",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/AdminPerfumeBasedFactorPerfumes"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "fullProduct": {
                "title": "اطلاعات محصول موجود برای ادمین",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string"
                    },
                    "volume": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "string"
                    },
                    "warranty": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "percent": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "string"
                    },
                    "discountEndTime": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "brand": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "searchAdmin": {
                "title": "محصولات فیلتر شده",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string"
                    },
                    "volume": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "string"
                    },
                    "warranty": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "percent": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "category name": {
                        "type": "string"
                    },
                    "brand name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AdminPerfumeBasedFactorPerfumes": {
                "title": "تمام محصولات ارایه دهنده یک محصول",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string"
                    },
                    "volume": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "string"
                    },
                    "warranty": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "string"
                    },
                    "sold": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "discountPercent": {
                        "type": "string"
                    },
                    "discountEndTime": {
                        "type": "string"
                    },
                    "discountCard": {
                        "type": "string"
                    },
                    "perfumeBasedFactor": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "volume": {
                                    "type": "string"
                                },
                                "price": {
                                    "type": "string"
                                },
                                "stock": {
                                    "type": "string"
                                },
                                "sold": {
                                    "type": "string"
                                },
                                "isActive": {
                                    "type": "string"
                                },
                                "warranty": {
                                    "type": "string"
                                },
                                "gender": {
                                    "type": "string"
                                },
                                "deleted": {
                                    "type": "string"
                                },
                                "updatedAt": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}